requirements.txt
streamlit
polars
google-genai
faker
duckdb